backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-limiter==3.5.0
python-dotenv==0.19.0
google-cloud-language==2.11.0
google-cloud-firestore==2.11.1
google-cloud-texttospeech==2.14.1
google-cloud-storage==2.13.0
openai==1.3.7
gunicorn==20.1.0
requests==2.31.0
pytest==7.4.3
pytest-mock==3.12.0

